# apps/backend/openapi/v1.yaml
...
openapi: 3.1.1
info:
  title: Argentum API
  # A date rather than a semver, matching what `GET /v1/me` reports as
  # `api_version` — "you are on 2026-07-30" locates a deploy, where "1.0.0"
  # locates nothing. A test pins this equal to `apiv1.Version`.
  version: '2026-07-30'
  summary: Ask for a report or an answer over HTTP.
  description: |
    Argentum turns a question about a company's own data into an answer or a
    branded document. This is the machine-facing half: the same agent the
    dashboard drives, reachable with an API key.

    ## The two products

    **Reports.** `POST /v1/reports/render` takes a document spec and returns a
    file — deterministic, sub-second, no LLM. `POST /v1/reports` takes a prompt
    and runs a real agent turn that ends by generating one.

    **Chat.** `POST /v1/chat` takes a question and returns an answer, either
    streamed over server-sent events or waited for on the connection.

    ## The contract every route inherits

    * **Authentication** is `Authorization: Bearer arg_…`. A dashboard session
      is refused here, and a key is refused on the dashboard's own `/api`
      routes: they are different authorities.
    * **Scopes** are fixed when a key is minted and cannot be changed. Each
      operation below names the one it requires under `x-argentum-scope`.
    * **Every response carries `X-Request-Id`**, including failures. Quote it
      in a support conversation.
...
      `error.message` is for a human.
...

    ## Streams

...

...

...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /v1/reports/render:
    post:
      operationId: renderReport
      tags: [Reports]
      summary: Render a document from a spec
      description: |
        Deterministic. A spec goes in, a file comes out, and nothing about it
        touches an LLM or a thread.

        **`Accept` decides the response body.** `application/json` (or an absent
        `Accept`) returns the document object with a short-lived presigned URL.
        The format's own content type — `application/pdf`, `text/csv`, the two
        OOXML types — or `application/octet-stream` returns the bytes inline,
        which is what a caller writing them straight to disk wants.

        A spec that outruns the synchronous window becomes a job and answers
        `202` with a report object, on the same shape `POST /v1/reports`
        returns. Write one collection path for both.
      x-argentum-scope: write:reports
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReportSpec' }
            # A real spec, small enough to read. The Postman collection is
            # generated from this value, so it is exercised every time someone
            # imports the collection and presses Send.
            example:
              spec_version: 2
              format: pdf
              title: Revenue by month
              locale: en
              currency: USD
              content:
                sections:
                  - type: heading
                    text: Revenue by month
                  - type: kpi_row
                    items:
                      - label: Total
                        value: { v: 270300, fmt: currency }
                        delta_pct: 10.5
                  - type: table
                    columns:
                      - { label: Month }
                      - { label: Revenue, fmt: currency }
                    rows:
                      - ['July', 128400]
                      - ['August', 141900]
                    total_row: [{ v: Total }, { v: 270300, fmt: currency }]
      responses:
        '200':
          description: |
            The rendered document. The body is JSON or the file's bytes,
            according to `Accept`.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
            Content-Disposition: { $ref: '#/components/headers/Content-Disposition' }
            X-Document-Id: { $ref: '#/components/headers/X-Document-Id' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RenderedDocument' }
            application/pdf:
              schema: { type: string, format: binary }
            text/csv:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
            application/vnd.openxmlformats-officedocument.presentationml.presentation:
              schema: { type: string, format: binary }
            application/octet-stream:
              schema: { type: string, format: binary }
        '202':
          description: |
            The spec outran the synchronous window and became a job. Poll
            `GET /v1/reports/{id}` or stream `GET /v1/reports/{id}/events`.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Report' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /v1/documents:
    get:
      operationId: listDocuments
      tags: [Documents]
      summary: List generated documents
      description: |
        Newest first, cursor-paginated.

        **No download URLs on the list.** Presigning is a signature per row, and
        a caller paging a hundred documents to find one is not about to fetch
        all hundred. `GET /v1/documents/{id}` issues the URL for the one they
        picked.
      x-argentum-scope: read:documents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: format
          in: query
          description: Only documents of this format.
          required: false
          schema: { $ref: '#/components/schemas/DocumentFormat' }
        - name: created_after
          in: query
          description: Only documents created at or after this RFC3339 timestamp.
          required: false
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          description: Only documents created at or before this RFC3339 timestamp.
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: One page of documents.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /v1/documents/{id}:
    get:
      operationId: getDocument
      tags: [Documents]
      summary: Fetch one document's metadata and a fresh download URL
      x-argentum-scope: read:documents
      parameters:
        - $ref: '#/components/parameters/DocumentID'
      responses:
        '200':
          description: |
            The document. `download_url` is absent when the object store could
            not be reached — the metadata is still true, and a caller listing
            documents is not necessarily about to download one.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer arg_<prefix>_<secret>`, minted in the dashboard
        under Settings → API Keys. The secret is shown once.

        **Header only.** The dashboard's own token may travel in a cookie or a
        query parameter because a browser cannot set a header on a WebSocket
        upgrade; neither applies to a machine caller, and both are how a
        credential ends up in an access log or a proxy trace.
  parameters:
    Limit:
      name: limit
      in: query
      description: |
        Page size. A value over the repository's cap is clamped rather than
        refused — a caller asking for 10 000 rows wants as many as they can get,
        and the cursor is what makes the rest reachable.
      required: false
      schema: { type: integer, minimum: 1, examples: [50] }
    Cursor:
      name: cursor
      in: query
      description: |
        The `next_cursor` from the previous page. Opaque: it encodes the
        `(created_at, id)` of the last row you saw, and hand-building one is a
        `400`, not a silent restart at page one.
      required: false
      schema: { type: string }
...
    DocumentID:
      name: id
      in: path
      required: true
      description: A document id.
      schema: { type: string, format: uuid }
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        A unique string per logical request; a UUID is the usual choice. Both
        SDKs generate one automatically when you do not supply it.

        A second request under the same key ends one of three ways:

        * the original finished — you get its result again, with
          `Idempotent-Replay: true`, and nothing is billed twice;
        * the original is still running — `409 request_in_flight`, carrying the
          ids of the work you are already waiting on;
        * the body differs from the original's — `409 idempotency_key_reuse`.
          Use a new key for a new request.

        A request that fails forgets its key, so a retry after a 500 is a fresh
        attempt. The one exception is `POST /v1/chat`'s 504, where the request
        failed and the work did not.
      schema: { type: string, maxLength: 255 }
...
  headers:
    X-Request-Id:
      description: |
        This request's id, on every response including failures. Send your own
        and it is echoed back, provided it is at most 128 characters of
        `[A-Za-z0-9_.:-]`.
      schema: { type: string, examples: ['req_1945841125adc2c7bb618224f5fa88fd'] }
    RateLimit-Limit:
      description: Requests permitted per minute on this key.
      schema: { type: integer }
    RateLimit-Remaining:
      description: Whole requests left in the bucket after this one.
      schema: { type: integer }
...
    Retry-After:
      description: Seconds to wait before retrying. Both SDKs honour it.
      schema: { type: integer }
    Idempotent-Replay:
      description: |
        `true` when this response replays an earlier request under the same
        `Idempotency-Key` rather than executing a new one.
      schema: { type: string, enum: ['true'] }
    Content-Disposition:
      description: '`attachment; filename="…"`, with the generated filename.'
      schema: { type: string }
    X-Document-Id:
      description: |
        The document's id, on the inline-bytes path. It is the only place the id
        appears when the body is a file rather than JSON.
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: |
        The request is malformed. `error.param` names the offending field
        whenever there is one to name.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: |
        No key, or one that is not usable. One message covers malformed,
        unknown, wrong-secret, revoked and expired: a caller holding a broken
        credential has no business learning which of the five it is.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: |
        A valid key without the scope this route requires. Scopes are fixed when
        a key is minted — mint a new one.

        On `POST /v1/chat` and `POST /v1/reports` it is also
        `agent_not_allowed`: the key is limited to named agents, and the turn
        would run as another. Name one of the key's agents in `agent_id`.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No such resource **for this company**. A cross-tenant id and a
        nonexistent id answer identically; anything else is an existence oracle
        over another company's data.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Two shapes, distinguished by `error.code`. `request_in_flight` carries
        `in_flight` with the ids of the work already running;
        `idempotency_key_reuse` means the same key arrived with a different
        body.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Idempotent-Replay: { $ref: '#/components/headers/Idempotent-Replay' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ConflictError' }
...
    TooLarge:
      description: The request body is over this deployment's cap.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests on this key.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
        RateLimit-Limit: { $ref: '#/components/headers/RateLimit-Limit' }
        RateLimit-Remaining: { $ref: '#/components/headers/RateLimit-Remaining' }
        RateLimit-Reset: { $ref: '#/components/headers/RateLimit-Reset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServerError:
      description: |
        Our fault. Quote `error.request_id` — it is the string that finds this
        attempt in our logs.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: |
        Either the whole public API is switched off (`api_disabled`, with
        `Retry-After`) or one subsystem this route needs is not configured on
        this deployment. The routes exist and are coming back, which is why this
        is not a 404.
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
        Retry-After: { $ref: '#/components/headers/Retry-After' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      description: |
        Every failure under `/v1` has this shape. A bare `{"error": "…"}` string
        anywhere here is a defect.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorDetail' }
      additionalProperties: false
...
    DocumentFormat:
      type: string
      description: |
        `mp4` is the same report spec as a silent 1080p video. It is rendered by
        a separate service and takes minutes rather than milliseconds, so every
        door that produces one is asynchronous: `POST /v1/reports/render` with
        `format: mp4` answers `202` whatever `Accept` says, and the file is
        collected the same three ways an agentic report is. It requires a
        report that makes an argument — a `kpi_row` or a `chart` — because a
        record such as an invoice cannot be scanned in a video. A deployment
        with no render service refuses it with `format_unavailable` and serves
        every other format as before.

        `carousel` is the same report spec as one to ten JPEG slides for a
        social feed, drawn by the same render service, and it follows every
        rule above: asynchronous, refused where there is no render service.
        The document is a zip of the pages, a `caption.txt` and a
        `carousel.json` manifest (caption, hashtags, one alt text per slide);
        the spec may carry `social.caption` (≤ 2 200 characters),
        `social.hashtags` (≤ 30) and `social.size`. A spec whose sections make
        more than ten slides is refused with the count, not truncated.

        **One slide is a whole post.** A promotion or a launch is one image:
        one `promo` section for a product with a price, or one `hero` section
        for anything else, and nothing else. Such a post is an announcement rather than an
        argument, so it is the one carousel exempt from the "argument about
        data rather than a record" rule above: a `hero` satisfies it where a
        `kpi_row` or a `chart` otherwise would. An invoice still cannot be a
        carousel, because a record carries neither.
      enum: [pdf, pptx, xlsx, csv, mp4, carousel]
    Document:
      type: object
      description: One generated file.
      required: [id, object, filename, format, size_bytes, source, created_at]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: document }
        filename: { type: string }
        format: { $ref: '#/components/schemas/DocumentFormat' }
        size_bytes: { type: integer }
        source:
          type: string
          description: |
            Which door produced it. `api` for either `/v1` report route, `agent`
            for the `generate_document` tool inside a dashboard turn.
          enum: [agent, api]
        thread_id:
          type: string
          format: uuid
          description: |
            Absent for a document the render door produced — it has no
            conversation behind it.
        created_at: { type: string, format: date-time }
        download_url:
          type: string
          format: uri
          description: |
            Presigned and short-lived, re-issued on every read rather than
            stored. Absent on list responses and when the object store could not
            be reached.
        expires_at: { type: string, format: date-time }
      additionalProperties: false
    RenderedDocument:
      type: object
      description: What the render door returns when `Accept` asks for JSON.
      required: [object, document]
      properties:
        object: { type: string, const: document }
        document: { $ref: '#/components/schemas/Document' }
        request_id: { type: string }
      additionalProperties: false
    DocumentPage:
      type: object
      required: [data, has_more]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Document' }
        has_more:
          type: boolean
          description: |
            Not `len(data) == limit`. A caller cannot tell a full last page from
            a full middle one, and asking for one more page to find out is a
            request nobody should have to make.
        next_cursor: { type: string }
      additionalProperties: false
...
    Report:
      type: object
      description: |
        One shape for both doors and for the poll route, so a caller writes the
        collection path once whether the job came from a prompt or from a spec
        that ran long.
      required: [id, object, status, kind, format, created_at]
      properties:
        id: { type: string, format: uuid }
        object: { type: string, const: report }
        status:
          type: string
          enum: [queued, running, completed, failed]
        kind:
          type: string
          description: '`agentic` for a prompt; `render` for a spec that outran the synchronous window.'
          enum: [agentic, render]
        format: { $ref: '#/components/schemas/DocumentFormat' }
        thread_id:
          type: string
          format: uuid
          description: |
            The conversation behind an agentic report — continue it through
            `POST /v1/chat`. Absent for a render job, which has none.
        document: { $ref: '#/components/schemas/Document' }
        error:
          type: string
          description: Why a `failed` report failed.
        created_at: { type: string, format: date-time }
        request_id: { type: string }
      additionalProperties: false
...
    ReportSpec:
      type: object
      description: |
        One file, one format, one content tree.

        `content` carries one of three shapes: `sections` for a PDF or a deck,
        `table` for a CSV, `sheets` or `table` for a workbook.
      required: [format, content]
      properties:
        spec_version:
          type: integer
          description: |
            `2` for the enterprise layout — cover page, running header, footer.
            Absent or `1` renders flat. A v1 spec is a v2 spec whose columns are
            all text, so this only decides whether the *renderer* offers the v2
            chrome: a document that has been rendering as a bare table for three
            months should not sprout a cover page because the backend was
            upgraded.
          enum: [1, 2]
        format: { $ref: '#/components/schemas/DocumentFormat' }
        filename:
          type: string
          description: Without a directory. Sanitised; the extension follows `format`.
        title: { type: string }
        locale:
          type: string
          description: '`id` or `en`. Empty derives it from `currency`.'
        currency:
          type: string
          description: An ISO 4217 code. Empty uses the company default.
        generated_at:
          type: string
          description: |
            RFC3339. It is a spec field rather than a clock read so the same
            spec renders the same bytes twice — pin it and the output is
            reproducible.
        meta: { $ref: '#/components/schemas/ReportSpecMeta' }
        content: { $ref: '#/components/schemas/ReportSpecContent' }
        social: { $ref: '#/components/schemas/ReportSpecSocial' }
      additionalProperties: false
...
    ReportSpecMeta:
      type: object
      description: |
        What lands in the PDF's info dictionary. A document that leaves the
        building with an empty Title in its properties is one a records system
        cannot file.
      properties:
        author: { type: string }
        subject: { type: string }
        keywords: { type: string }
      additionalProperties: false
    ReportSpecContent:
      type: object
      properties:
        sections:
          type: array
          items: { $ref: '#/components/schemas/ReportSpecSection' }
        table: { $ref: '#/components/schemas/ReportSpecTable' }
        sheets:
          type: array
          items: { $ref: '#/components/schemas/ReportSpecSheet' }
      additionalProperties: false
...